Update c.py
Clicks on <most> items of each page, slowly iterates through pages and cities because of all the .sleep statements. All sleep and wait methods are highly variable in the amount of time needed but are necessary for now. Prints the number of items clicked on as they are visited just for visualization purposes. Break statements are ugly, also likely to change