Skip scanned pages
If ALL items on a page already have folders in the data directory, the entire page will be skipped. If there is AT LEAST ONE item on the page (120 items), all 120 will be visited. This could be changed in the future to visit the first item chronologically that hasn't been scraped but will require some more logic. The worst-case scenario that 119 items are revisited is much less taxing than the implementation where no pages were skipped at all, so for now this is acceptable.